refactor: pages dir updated, routes dir criated
